Move editor to home view
