+++ package.json
@@ -3,9 +3,12 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "aws-amplify": "^2.2.1",
     "react": "^16.10.2",
+    "react-bootstrap": "^1.0.0-beta.16",
     "react-dom": "^16.10.2",
     "react-quill": "^1.3.1",
+    "react-router-dom": "^5.1.2",
     "react-scripts": "^3.3.0"
   },
   "scripts": {
@@ -22,6 +25,7 @@
     "@types/node": "^12.7.12",
     "@types/react": "^16.9.5",
     "@types/react-dom": "^16.0.7",
+    "@types/react-router-dom": "^5.1.3",
     "typescript": "^3.6.4"
   },
   "browserslist": {
