+++ public/index.html
@@ -19,6 +19,7 @@
   <!-- <link rel="manifest" href="%PUBLIC_URL%/manifest.json">
   <link rel="shortcut icon" href="%PUBLIC_URL%/favicon.ico"> -->
   <!-- <link rel="stylesheet" href="https://cdn.quilljs.com/1.2.6/quill.snow.css"> -->
+  <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css">
   <link rel="stylesheet" href="node_modules/react-quill/dist/quill.snow.css">
   <link rel="stylesheet" href="node_modules/react-quill/dist/quill.bubble.css">
   <link rel="stylesheet" href="node_modules/react-quill/dist/quill.core.css">
@@ -31,7 +32,7 @@
       work correctly both with client-side routing and a non-root public URL.
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
-  <title>React App</title>
+  <title>Logd Note-Taking App</title>
 </head>
 
 <body>
