+++ src/components/Routes.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+import { Route, Switch } from "react-router-dom";
+import Home from "./Home";
+
+export default function Routes() {
+  return (
+    <Switch>
+      <Route path="/" exact component={Home} />
+    </Switch>
+  );
+}
