+++ src/index.tsx
@@ -3,6 +3,37 @@ import * as ReactDOM from "react-dom";
 import App from "./components/App";
 import "./styles/styles.css";
 import registerServiceWorker from "./registerServiceWorker";
+import Amplify from 'aws-amplify';
+import config from './config';
+import { BrowserRouter as Router } from 'react-router-dom';
 
-ReactDOM.render(<App />, document.getElementById("root") as HTMLElement);
+Amplify.configure({
+    Auth: {
+      mandatorySignIn: true,
+      region: config.cognito.REGION,
+      userPoolId: config.cognito.USER_POOL_ID,
+    //   identityPoolId: config.cognito.IDENTITY_POOL_ID,
+      userPoolWebClientId: config.cognito.APP_CLIENT_ID
+    },
+    // Storage: {
+    //   region: config.s3.REGION,
+    //   bucket: config.s3.BUCKET,
+    //   identityPoolId: config.cognito.IDENTITY_POOL_ID
+    // },
+    // API: {
+    //   endpoints: [
+    //     {
+    //       name: "notes",
+    //       endpoint: config.apiGateway.URL,
+    //       region: config.apiGateway.REGION
+    //     },
+    //   ]
+    // }
+  });
+
+ReactDOM.render(
+    <Router>
+    <App />
+  </Router>,
+document.getElementById("root") as HTMLElement);
 registerServiceWorker();
